+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Handling of missing GitRepoTemplates ([#110])
+### Added
+- Define cluster template for tenant ([#110])
+
 ## [v0.2.2] - 2020-09-10
 ### Fixed
 - Update latest image tag in deployment ([#104])
@@ -100,3 +106,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#100]: https://github.com/projectsyn/lieutenant-operator/pull/100
 [#103]: https://github.com/projectsyn/lieutenant-operator/pull/103
 [#104]: https://github.com/projectsyn/lieutenant-operator/pull/104
+[#110]: https://github.com/projectsyn/lieutenant-operator/pull/110